requirements.txt
BeautifulSoup==3.2.1
Django==1.5.2
Pillow==2.6.1
South==0.8.2
appy==0.8.5
argparse==1.2.1
chardet==2.1.1
cmsplugin-htmlsitemap==0.1.3
diff-match-patch==20120106
django-ace==0.3.1
django-admin-tools==0.5.1
django-adv-cache-tag==0.1.1
django-appconf==0.6
django-cached-authentication-middleware==0.2.0
django-classy-tags==0.4
django-compressor==1.3
django-extensions==1.1.1
django-generic-positions==0.1.4
django-modeltranslation==0.6.1
django-mptt==0.5.2
django-object-tools==0.0.7
django-reset==0.2.0
django-reversion==1.7.1
django-rosetta==0.7.2
django-sekizai==0.7
django-templateaddons==0.1
django-tinymce==1.5
html5lib==1.0b3
importers==0.1.1
jsonfield==0.9.17
lxml==3.4.1
newrelic==1.13.1.31
pilkit==1.1.3
polib==0.7.0
psycopg2==2.5.4
python-magic==0.4.3
python-memcached==1.53
pytz==2014.10
six==1.3.0
solid-i18n==0.5.1
stevedore==0.10
tablib==0.9.11
wsgiref==0.1.2
